Guild: fix swapped event-log fields, a reload narrowing, and stale rank definitions
EventType and NewRank were the wrong way round in SMSG_GUILD_EVENT_LOG. The
consumer settles it: GetGuildEventInfo switches on the byte from +20 to yield
invite/join/promote/demote/remove/quit and passes +21 to the rank-name lookup.
The layout now lives in an inline builder with a byte-exact fixture using
distinct values, since equal ones could not have caught the swap.

Guild bank tab rights and slots reloaded through GetUInt16 while the column is
int unsigned and the client permits 100000, folding it to 34464 on every
restart. Registering SET_RANK made that path live.

Replaces the rank-0 refusal, which broke legitimate guildmaster renaming, with
an OR of GR_RIGHT_ALL matching the load path so extension bits survive.

Rank definitions now broadcast from all four mutating handlers via
BroadcastRankDefinitions. Roster alone carries only rank IDs, so other members
kept stale names, rights and limits; reordering was worst, since SwitchRank
swaps both definitions and member rank IDs. The AddRank and DelRank half was
written by Codex after it raised the finding, and reviewed by GLM.

+++ src/game/Object/GuildBank.cpp
@@ -671,8 +671,10 @@ bool Guild::LoadBankRightsFromDB(QueryResult* guildBankTabRightsResult)
             break;
         uint8 TabId        = fields[1].GetUInt8();
         uint32 rankId      = fields[2].GetUInt32();
-        uint16 right       = fields[3].GetUInt16();
-        uint16 SlotPerDay  = fields[4].GetUInt16();
+        // 32-bit: the column is int unsigned and the client permits values up to
+        // 100000, which a 16-bit read silently folded to 34464 on every restart.
+        uint32 right       = fields[3].GetUInt32();
+        uint32 SlotPerDay  = fields[4].GetUInt32();
 
         SetBankRightsAndSlots(rankId, TabId, right, SlotPerDay, false);
     }
